content_service/requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.35.1
azure-mgmt-core==1.6.0
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
clerk-backend-api
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-ipware==7.0.1
django-mongodb-backend==5.2.0b1
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docopt==0.6.2
drf-spectacular==0.28.0
flower==2.0.1
frozenlist==1.7.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.4
Markdown==3.8.2
mongo_migrate==0.1.2
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.4.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.2.1
pycparser==2.22
pydantic-core==2.41.5
PyJWT==2.9.0
pymongo==4.13.2
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-ipware==3.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
svix==1.81.0
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
yarg==0.1.10
yarl==1.20.1
